Add tests base and one more solution
