+++ source/leetcode/counter/counter-increment-then-return.js
