+++ source/tests/counter/counter-increment-then-return.test.js
@@ -1,0 +1,4 @@
+const counterFunction = require("../../leetcode/counter/counter-increment-then-return");
+const testCounterFunction = require("./counter-tests-base");
+
+testCounterFunction(counterFunction);
