+++ source/tests/counter/counter-postfix-increment.test.js
@@ -1,0 +1,4 @@
+const counterFunction = require("../../leetcode/counter/counter-postfix-increment");
+const testCounterFunction = require("./counter-tests-base");
+
+testCounterFunction(counterFunction);
